+++ packages/sdk/src/index.ts
@@ -30,7 +30,6 @@ export function createRaribleSdk(
 		createTezosSdk(filterWallet(wallet, "TEZOS"), apis, config.tezosNetwork)
 	)
 
-
 	return {
 		...instance,
 		nft: {