+++ packages/sdk/src/sdk-blockchains/tezos/cancel.test.ts
@@ -1,9 +1,7 @@
 // eslint-disable-next-line camelcase
 import { in_memory_provider } from "tezos-sdk-module/dist/providers/in_memory/in_memory_provider"
 import { TezosWallet } from "@rarible/sdk-wallet"
-import { toContractAddress, toOrderId } from "@rarible/types"
-import BigNumber from "bignumber.js"
-import { deploy_nft_public } from "tezos-sdk-module"
+import { toContractAddress } from "@rarible/types"
 import { createRaribleSdk } from "../../index"
 import { MintType } from "../../types/nft/mint/domain"
 import { delay, retry } from "../../common/retry"
@@ -17,31 +15,8 @@ describe("cancel test", () => {
 	const sdk = createRaribleSdk(wallet, "dev")
 
 	let nftContract: string = "KT1CHLDcbogVfVtRbg2TZKvL5p5w9WvhYe2G"
-	// let nftContract: string = "KT1Co8iAMMSj7aC8z5Ruk7m3THbiFcAjhH1o"
 
-	/*
-	beforeAll(async () => {
-		const provider = {
-			tezos,
-			api: "https://rarible-api.functori.com/v0.1",
-			config: {
-				exchange: "KT1AguExF32Z9UEKzD5nuixNmqrNs1jBKPT8",
-				fees: new BigNumber(0),
-				nft_public: "",
-				mt_public: "",
-			},
-		}
-		const sender = await tezos.address()
-		console.log("sender", sender)
-		const nft = await deploy_nft_public(provider, sender)
-		console.log("nft", nft.contract)
-		nftContract = nft.contract as any
-		await nft.confirmation()
-	})
-
-   */
-
-	test("cancel order", async () => {
+	test.skip("cancel order", async () => {
 		const mintResponse = await sdk.nft.mint({
 			collectionId: toContractAddress(`TEZOS:${nftContract}`),
 		})
@@ -87,7 +62,6 @@ describe("cancel test", () => {
 		await cancelTx.wait()
 		await retry(10, 1000, async () => {
 			const canceledOrder = await sdk.apis.order.getOrderById({
-				// id: toOrderId("TEZOS:fa33e0e6acad0c787b76ba9c387fc1c73cc8b27859bf1667c53535546807543e"),
 				id: orderId,
 			})
 			if (canceledOrder.status !== "CANCELLED") {