+++ packages/sdk/src/sdk-blockchains/tezos/fill.ts
@@ -1,11 +1,9 @@
 import type {
-	AssetType as TezosAssetType, Provider,
-	Asset as TezosLibAsset,
+	Provider,
 } from "tezos-sdk-module/dist/common/base"
 import { Action } from "@rarible/action"
-import type { AssetType, Order } from "@rarible/api-client"
 import type {
-	OrderForm, Part, Part as TezosPart,
+	Part, Part as TezosPart,
 } from "tezos-sdk-module/dist/order"
 // eslint-disable-next-line camelcase
 import { fill_order, get_address } from "tezos-sdk-module"
@@ -19,13 +17,9 @@ import {
 import { BlockchainTezosTransaction } from "@rarible/sdk-transaction"
 import type {
 	Order as TezosOrder,
-	Asset as TezosClientAsset,
 } from "tezos-api-client"
 import BigNumber from "bignumber.js"
 import type { TezosProvider } from "tezos-sdk-module/dist/common/base"
-// eslint-disable-next-line camelcase
-import { get_public_key } from "tezos-sdk-module/dist/common/base"
-import { order_of_json } from "tezos-sdk-module/dist/order"
 import type {
 	FillRequest,
 	PrepareFillRequest,
@@ -37,7 +31,6 @@ import {
 } from "../../types/order/fill/domain"
 import type { ITezosAPI, MaybeProvider, PreparedOrder } from "./common"
 import { convertOrderToFillOrder, covertToLibAsset, isExistedTezosProvider } from "./common"
-
 
 export class TezosFill {
 	constructor(
@@ -59,6 +52,7 @@ export class TezosFill {
 			type: "RARIBLE_V2",
 			maker: order.maker,
 			maker_edpk: order.makerEdpk,
+			taker_edpk: order.takerEdpk,
 			make: covertToLibAsset(order.make),
 			take: covertToLibAsset(order.take),
 			salt: order.salt,
@@ -81,8 +75,8 @@ export class TezosFill {
 		})) || []
 	}
 
-	// async getPreparedOrder(request: PrepareFillRequest): Promise<PreparedOrder> {
-	async getPreparedOrder(request: PrepareFillRequest): Promise<OrderForm> {
+	async getPreparedOrder(request: PrepareFillRequest): Promise<PreparedOrder> {
+	// async getPreparedOrder(request: PrepareFillRequest): Promise<OrderForm> {
 		if ("order" in request) {
 			return convertOrderToFillOrder(request.order)
 		} else if ("orderId" in request) {
@@ -93,8 +87,7 @@ export class TezosFill {
 			const order = await this.apis.order.getOrderByHash({
 				hash,
 			})
-			// return this.convertTezosOrderToForm(order)
-			return order_of_json(order)
+			return this.convertTezosOrderToForm(order)
 		} else {
 			throw new Error("Request error")
 		}
@@ -131,9 +124,7 @@ export class TezosFill {
 					payouts: this.convertOrderPayout(fillRequest.payouts),
 					origin_fees: this.convertOrderPayout(fillRequest.originFees),
 					infinite: fillRequest.infiniteApproval,
-					// edpk: await get_public_key(provider),
 				}
-				console.log("req", JSON.stringify(preparedOrder, null, "  "), JSON.stringify(request, null, " "))
 				const fillResponse = await fill_order(
 					provider,
 					preparedOrder,
@@ -144,10 +135,8 @@ export class TezosFill {
 		})
 
 		return {
-			// multiple: this.isMultiple(preparedOrder),
-			multiple: true,
-			// maxAmount: await this.getMaxAmount(preparedOrder),
-			maxAmount: toBigNumber("1"),
+			multiple: this.isMultiple(preparedOrder),
+			maxAmount: await this.getMaxAmount(preparedOrder),
 			baseFee: parseInt(provider.config.fees.toString()),
 			originFeeSupport: OriginFeeSupport.FULL,
 			payoutsSupport: PayoutsSupport.MULTIPLE,