+++ packages/sdk/src/sdk-blockchains/tezos/test/await-for-item-supply.ts
@@ -3,8 +3,10 @@ import type BigNumber from "bignumber.js"
 import { retry } from "../../../common/retry"
 import type { IRaribleSdk } from "../../../domain"
 
-export async function awaitForItemSupply(sdk: IRaribleSdk, itemId: ItemId, supply: string | number | BigNumber) {
-	await retry(10, 1000, async () => {
+export async function awaitForItemSupply(
+	sdk: IRaribleSdk, itemId: ItemId, supply: string | number | BigNumber
+): Promise<string> {
+	return retry(10, 1000, async () => {
 		const item = await sdk.apis.item.getItemById({
 			itemId,
 		})
@@ -13,5 +15,6 @@ export async function awaitForItemSupply(sdk: IRaribleSdk, itemId: ItemId, suppl
 		if (itemSupply !== requireSupply) {
 			throw new Error(`Expected supply ${requireSupply}, but current supply ${itemSupply}`)
 		}
+		return itemSupply
 	})
 }