+++ packages/sdk/src/sdk-blockchains/tezos/token-id.ts
@@ -1,0 +1,31 @@
+import type { TezosProvider } from "tezos-sdk-module/dist/common/base"
+// eslint-disable-next-line camelcase
+import { get_next_token_id } from "tezos-sdk-module"
+import { toBinary } from "@rarible/types"
+import type { GenerateTokenIdRequest, TokenId } from "../../types/nft/generate-token-id"
+import { getRequiredProvider, getTezosAddress } from "./common"
+import type { ITezosAPI, MaybeProvider } from "./common"
+
+export class TezosTokenId {
+	constructor(
+		private provider: MaybeProvider<TezosProvider>,
+		private apis: ITezosAPI,
+	) {
+		this.generateTokenId = this.generateTokenId.bind(this)
+	}
+
+	async generateTokenId({ collection }: GenerateTokenIdRequest): Promise<TokenId> {
+		const tokenId = await get_next_token_id(
+			getRequiredProvider(this.provider),
+			getTezosAddress(collection)
+		)
+		return {
+			tokenId: tokenId.toString(),
+			signature: {
+				v: 0,
+				r: toBinary("0"),
+				s: toBinary("0"),
+			},
+		}
+	}
+}